Send memory usage information together with metrics
We special case Linux since getting accurate memory usage values on it
is most important. On macOS the memory value returned by `os.freemem()`
is also inaccurate for similar reasons than on Linux, but it would be
quite rare rare to find it as a production web server's OS. Not sure how
accurate `os.freemem()` is on other platforms that Node.js supports
either, but should be better than nothing.

Good to note that `MemAvailable` in `/proc/meminfo` is only available on
Linux kernel 3.14 and up, but since 3.14 has already reached end of
life we can pretty safely ignore support for systems before it.